refactor(cli): extract shared QueryResultFormatter utility
Extract ~100 lines of duplicated code into shared utilities:

- Create QueryResultFormatter in Infrastructure/Output with:
  - WriteTableOutput() - render query results as aligned table
  - WriteCsvOutput() - render query results as CSV
  - TruncateValue() - truncate long cell values with ellipsis
  - EscapeCsvField() - escape CSV special characters

- Create shared HistoryEntryOutput model used by GetCommand and
  ListCommand for JSON output

- Update SqlCommand and ExecuteCommand to use QueryResultFormatter
- Update GetCommand and ListCommand to use shared HistoryEntryOutput

Closes #381
